git: let clang generate pdbs through ldflags
We can ask the LLD linker to generate PDBs for us. This commit
makes LDFLAGS configurable and sets the configuration for the Clang
toolchain. As some of the builtins are hard linked to git-remote-http,
LLD doesn't generate PDBs for them. By hard linking them ourselves, we
ensure that mingw-w64-git-pdb can be built correctly.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>