git: try, again, to fix the path of `cv2pdb-strip`
It is a bit tricky, apparently. If the `src/git` directory exists
already, at the point where `STRIP` is defined, the value of `$PWD`
evaluates to the directory where the `PKGBUILD` file is located.

However, if it does not exist, it will be created, and then `$PWD` will
evaluate to the `src/git/` subdirectory.

Let's use a Bash-ism to determine the location of the `PKGBUILD` file
and construct a path relative to said location. This should always work:
`makepkg` relies on many a Bash-ism already (e.g. Bash arrays).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>