curl: align with MSYS2
The MSYS2 project, from which Git for Windows' development environment
is derived, switched to building three flavors of libcurl a long time
ago: the OpenSSL flavor, the GNUTLS one, and the Secure Channel
("winssl") one.

To reduce maintenance burden, and also for the benefit that it will be
easier to upgrade OpenSSL from v1.1 to v3 in Git for Windows' SDK, let's
imitate that strategy.

Concretely, these are the changes that are made to align better with
MSYS2's build definition:

- The `pathtools.[ch]` files are copied directly, and kept in sync with
  `mingw-w64-pathtools`' copies.

- Three flavors are built: mingw-w64-curl, mingw-w64-curl-gnutls and
  mingw-w64-curl-winssl.

The following changes are Git for Windows-only and intentionally deviate
from MSYS2:

- We do not build static versions of libcurl.

- We still build `-pdb` packages.

- The `ca-bundle.crt` file needs to be found even if the DLL is in Git's
  `libexec/` directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>