mingw-w64-git: swap `windir` for `SystemRoot`
Currently Git CMD looks for `doskey.exe`, its `git` alias enabler, in
`%windir%\system32`. `windir`, as a relic of times predating Windows NT
is a system environment variable defined in Session Manager's variable
store that, being regular & writable, can be freely altered system-wide.
It has been superseded by `SystemRoot` (it points to it anyway).

As `SystemRoot` is a boot configuration parameter and thus predefined,
it is more resilient to attempts of tampering with environment variables.

Signed-off-by: Mateusz Kazimierczuk <mataha@shimiko.download>