clang: build a smaller clang/llvm
Msys2 builds pretty versatile Clang and LLVM packages with LLVM backends
to cross compile for various target architectures. That comes at the
cost of big binaries, that challenge how we manage the Git for Windows
SDK[1]. But for Git for Windows we don't need all of this. We only need
a Clang that can compile ARM64 binaries on ARM64.

[1] https://github.com/git-for-windows/git-sdk-arm64/issues/7

Signed-off-by: Matthias Aßhauer <mha1993@live.de>