pcre2: work around incorrect assumptions in the test suite
The tests verify that a certain amount of memory is allocated for given
patterns. The way those tests are implemented, these sizes are
hardcoded, and they work well as long as the code is compiled using
a 64-bit CPU architecture.

However, when compiling for i686, all of those numbers are off by 24.

Work around this for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>