ci: implement a (temporary) workaround to help switching to `asciidoctor`
We want to switch from using `asciidoctor-extensions` to using regular
`asciidoctor` instead (the custom macros are no longer needed).

To that end, we want to install `mingw-w64-asciidoctor`, but that would fail due
to a conflict with the `mingw-w64-asciidoctor-extensions` package:

  error: failed to commit transaction (conflicting files)
  mingw-w64-x86_64-asciidoctor: /mingw64/bin/asciidoctor exists in filesystem

Let's work around that by explicitly removing the latter package as
needed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>